pyproject.toml
[tool.poetry]
name = "metalib"
version = "0.1.0"
description = ""
authors = ["thaeber"]

[tool.poetry.dependencies]
python = ">=3.7"
pandas = "^1.1.3"
toolz = "^0.11.1"
pyyaml = "^5.3.1"
"ruamel.yaml" = "^0.16.12"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
yapf = "^0.30.0"
pylint = "^2.6.0"
bump2version = "^1.0.1"
mypy = "^0.790"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
